Use delegate * for other languages
