+++ mdoc/Mono.Documentation/Updater/Formatters/CSharpFullMemberFormatter.cs
@@ -562,67 +562,6 @@ namespace Mono.Documentation.Updater.Formatters
             return base.AppendRequiredModifierTypeName(buf, type, context);
         }
 
-        protected override void AppendFunctionPointerTypeName(
-            StringBuilder buf, FunctionPointerType type, IAttributeParserContext context)
-        {
-            buf.Append("delegate*");
-
-            var callingConvention = GetCallingConvention(type);
-            if (callingConvention != MethodCallingConvention.Default.ToString())
-            {
-                buf.Append(" unmanaged");
-                if (!string.IsNullOrEmpty(callingConvention))
-                {
-                    buf.Append("[").Append(callingConvention).Append("]");
-                }
-            }
-
-            buf.Append("<");
-            if (type.Parameters?.Count > 0)
-            {
-                AppendParameters(buf, type.Parameters);
-                buf.Append(", ");
-            }
-            AppendReturnTypeName(buf, type, true);
-            buf.Append(">");
-        }
-
-        private string GetCallingConvention(FunctionPointerType type)
-        {
-            var callingConvention = type.CallingConvention.ToString("D");
-            // Cecil lib uses "9" to stands for "Unmanaged Ext"
-            if (callingConvention != "9")
-            {
-                return NormalizeCallingConvention(type.CallingConvention);
-            }
-            else
-            {
-                StringBuilder buf = new StringBuilder();
-                AssembleCallingConvention(type.ReturnType, buf);
-                return buf.ToString();
-            }
-        }
-
-        private string NormalizeCallingConvention(MethodCallingConvention callingConvention)
-        {
-            if (callingConvention == MethodCallingConvention.C) return "Cdecl";
-            var callConv = callingConvention.ToString().ToLower();
-            return char.ToUpper(callConv[0]) + callConv.Substring(1);
-        }
-
-        private void AssembleCallingConvention(TypeReference type, StringBuilder buf)
-        {
-            if (!(type is OptionalModifierType optionalModifierType)) return;
-
-            var modifier = optionalModifierType.ModifierType.FullName;
-            if (modifier.StartsWith(Consts.CallConvPrefix))
-            {
-                if (!string.IsNullOrEmpty(buf.ToString())) buf.Append(", ");
-                buf.Append(modifier.Substring(Consts.CallConvPrefix.Length));
-                AssembleCallingConvention(optionalModifierType.ElementType, buf);
-            }           
-        }
-
         protected override StringBuilder AppendGenericMethod (StringBuilder buf, MethodDefinition method)
         {
             if (method.IsGenericMethod ())
@@ -653,24 +592,18 @@ namespace Mono.Documentation.Updater.Formatters
             {
                 if (DocUtils.IsExtensionMethod (method))
                     buf.Append ("this ");
-                AppendParameters(buf, parameters);
+                AppendParameter(buf, parameters[0]);
+                for (int i = 1; i < parameters.Count; ++i)
+                {
+                    buf.Append(", ");
+                    AppendParameter(buf, parameters[i]);
+                }
             }
 
             return buf.Append (end);
         }
 
-        private void AppendParameters(StringBuilder buf, IList<ParameterDefinition> parameters)
-        {
-            if (parameters == null || parameters.Count == 0) return; 
-            AppendParameter(buf, parameters[0]);
-            for (int i = 1; i < parameters.Count; ++i)
-            {
-                buf.Append(", ");
-                AppendParameter(buf, parameters[i]);
-            }
-        }
-
-        private StringBuilder AppendParameter (StringBuilder buf, ParameterDefinition parameter)
+        protected override StringBuilder AppendParameter(StringBuilder buf, ParameterDefinition parameter)
         {
             TypeReference parameterType = parameter.ParameterType;
             var refType = new BitArray(3);
